Moved Truss_Plotter to a separate file. Test for Truss_Plotter Config are working.
